+++ seed/fastapi/enum/resources/inlined_request/service/send_enum_inlined_request.py
@@ -13,7 +13,7 @@ except ImportError:
 
 
 class SendEnumInlinedRequest(pydantic.BaseModel):
-    value: typing.Optional[Operand]
+    operand: typing.Optional[Operand]
 
     def json(self, **kwargs: typing.Any) -> str:
         kwargs_with_defaults: typing.Any = {"by_alias": True, "exclude_unset": True, **kwargs}