+++ seed/python-sdk/object/src/seed/core/__init__.py
@@ -3,7 +3,7 @@
 from .api_error import ApiError
 from .client_wrapper import AsyncClientWrapper, BaseClientWrapper, SyncClientWrapper
 from .datetime_utils import serialize_datetime
-from .file import FileTypes, convert_file_dict_to_httpx_tuples
+from .file import File, convert_file_dict_to_httpx_tuples
 from .jsonable_encoder import jsonable_encoder
 from .remove_none_from_dict import remove_none_from_dict
 from .request_options import RequestOptions
@@ -12,7 +12,7 @@ __all__ = [
     "ApiError",
     "AsyncClientWrapper",
     "BaseClientWrapper",
-    "FileTypes",
+    "File",
     "RequestOptions",
     "SyncClientWrapper",
     "convert_file_dict_to_httpx_tuples",