+++ seed/ts-sdk/literal/src/api/resources/index.ts
@@ -1,3 +1,9 @@
-export * as literal from "./literal";
-export * from "./literal/types";
-export * from "./literal/client/requests";
+export * as reference from "./reference";
+export * from "./reference/types";
+export * as headers from "./headers";
+export * as inlined from "./inlined";
+export * as path from "./path";
+export * as query from "./query";
+export * from "./headers/client/requests";
+export * from "./inlined/client/requests";
+export * from "./query/client/requests";